fix ItadakiDataEdict build failure on Windows
This project always failed on Windows (or at least on my machine)
because the call to `File.renameTo(File)` always failed, preventing
the final dictionary file from being created. Fix this simply copying
and deleting when the `renameTo` method fails. This can still possibly
fail and of course takes longer, but is not as buggy. Note in STDERR
that this fallback is being used. This could probably be more easily
resolved with Java NIO, but we are supporting Java 1.5.

Also make changes to make this easier to debug in the future: fail the
ant build when `java` exits with an error, throw a separate exception
if an old index file could not be deleted, and note in STDERR if the
temp file could not be removed. Delete the temporary index file in
the ant `clean` target in case it's still around.