fix properties storage on Windows
The dictionary properties could not be stored on Windows because the
path logic created something like
`C:/users/name/C:/users/name/application data". Fix this so that the
file can be properly written on Windows.

As there are known problems with getting the home and application data
folders in a cross-platform manner, the class should probably be
changed to use the Preferences API instead.