Redo Gradle build for edict download
Change the versioning method, as well. Since edict is updated everyday and
the only available version is the date printed inside of it, just use today's
date as the version number. Technically they might not match due to a
timezone difference, but hopefully that won't be a big problem for anyone.